requirements.txt
# Core libraries
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
pillow==9.5.0
scikit-learn==1.2.2

# PyTorch and related libraries
torch==2.1.0  # Recommended upgrade
torchvision==0.16.0  # Recommended upgrade

# Progress bar
tqdm==4.65.0

# API and Web Framework
fastapi==0.115.1
uvicorn==0.29.0
pydantic==2.10.6
requests==2.32.3
